Open "Pas de bénéfices" link in new window
Opening the link "ne cherchez pas à faire de bénéfices" go to a help page, but replace the current page location. So if you click on the link, all the data you entered in the form may be lost. I added a "target=_blank" in the <a> tag to open the page in a new window and avoid the form data loss.